Move LanguageContext under Header folder
